Increased the size of a mapped_read->num_mappings to prevent overflows.

+++ src/mapped_read.h
@@ -266,7 +266,7 @@ set_cond_prob_in_mapped_read_location(
 
 struct mapped_read_t {
     unsigned long read_id;
-    unsigned short num_mappings;
+    unsigned int num_mappings;
     /* the database that this read is in - useful because the 
        DB often conatains meta data ( ie, frag len dist ) */
     struct mapped_reads_db* rdb;
